--- a/epilogue-runtime/src/main/java/edu/wpi/first/epilogue/logging/NestedBackend.java
+++ b/epilogue-runtime/src/main/java/edu/wpi/first/epilogue/logging/NestedBackend.java
@@ -17,6 +17,15 @@ public class NestedBackend implements EpilogueBackend {
   private final EpilogueBackend m_impl;
   private final Map<String, NestedBackend> m_nestedBackends = new HashMap<>();
 
+  // String concatenation can be expensive, especially for deeply nested hierarchies with many
+  // logged fields. For example, logging a hypothetical Robot.elevator.io.getHeight() would result
+  // in "/Robot/" + "elevator/" + "io/" + "getHeight"; three concatenations and string and byte
+  // array allocations that need to be cleaned up by the GC. Caching the results means those
+  // allocations only occur once, resulting in no GC (the strings are always referenced in the
+  // cache), and minimal time costs (the String object caches its own hash code, so all we do is an
+  // O(1) table lookup per concatenation)
+  private final Map<String, String> m_prefixedIdentifiers = new HashMap<>();
+
   /**
    * Creates a new nested backed underneath another backend.
    *
@@ -33,6 +42,26 @@ public class NestedBackend implements EpilogueBackend {
     this.m_impl = impl;
   }
 
+  /**
+   * Fast lookup to avoid redundant `m_prefix + identifier` concatenations. If the identifier has
+   * not been seen before, we compute the concatenation and cache the result for later invocations
+   * to read. This avoids redundantly recomputing the same concatenations every loop and
+   * significantly cuts down on the CPU and memory overhead of the Epilogue library.
+   *
+   * @param identifier The identifier to prepend with {@link #m_prefix}.
+   * @return The concatenated string.
+   */
+  private String withPrefix(String identifier) {
+    // Using computeIfAbsent would result in a new lambda object allocation on every call
+    if (m_prefixedIdentifiers.containsKey(identifier)) {
+      return m_prefixedIdentifiers.get(identifier);
+    }
+
+    String result = m_prefix + identifier;
+    m_prefixedIdentifiers.put(identifier, result);
+    return result;
+  }
+
   @Override
   public EpilogueBackend getNested(String path) {
     return m_nestedBackends.computeIfAbsent(path, k -> new NestedBackend(k, this));
@@ -40,76 +69,76 @@ public class NestedBackend implements EpilogueBackend {
 
   @Override
   public void log(String identifier, int value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, long value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, float value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, double value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, boolean value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, byte[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, int[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, long[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, float[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, double[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, boolean[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, String value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public void log(String identifier, String[] value) {
-    m_impl.log(m_prefix + identifier, value);
+    m_impl.log(withPrefix(identifier), value);
   }
 
   @Override
   public <S> void log(String identifier, S value, Struct<S> struct) {
-    m_impl.log(m_prefix + identifier, value, struct);
+    m_impl.log(withPrefix(identifier), value, struct);
   }
 
   @Override
   public <S> void log(String identifier, S[] value, Struct<S> struct) {
-    m_impl.log(m_prefix + identifier, value, struct);
+    m_impl.log(withPrefix(identifier), value, struct);
   }
 }