Closes #8189

Data was collected with JFR and AdvantageScope using a real (and fairly small) robot program running in simulation. The program was running at a 200Hz loop rate on Java 21 with ZGC and a 512MB max heap size. Larger programs with more logged fields will see better performance improvements

Key changes:

- Caching path concatenations in `NestedBackend` avoids huge numbers of memory allocations stemming from byte array and string object instantiations.
- Replacing lambdas with imperative equivalents (for lambdas with local captures) or caching in class members (for lambdas without local captures) slightly improves runtime performance and also saves several megabytes of object allocations, helping with GC pressure.

Results:
- Saved 2.73GiB of memory allocations per minute on byte arrays (~258KiB per loop)
- Saved 0.450GiB of memory allocations per minute on String objects (~39KiB per loop)
- Saved 30MiB of memory allocations per minute on lambda objects (~2.6KiB per loop)
- Median logging time per loop improved from 0.459ms to 0.374ms (a ~1.2x speedup)
- 90th %ile dropped from 0.787ms to 0.612ms
- 95th %ile dropped from 1.083ms to 0.810ms
- (99th %ile excluded due to outliers at program startup)